Add singlepose detection models and EfficientDetLite0 (#442) (#443)
* Add SSDMobileNetv1 model to ONNXModels

* Add EfficientDetLite0 model

* Add MovenetLighting

* Add MovenetThunder
