+++ onnx/src/commonMain/kotlin/org/jetbrains/kotlinx/dl/api/inference/onnx/objectdetection/ObjectDetectionModelBase.kt
@@ -68,7 +68,7 @@ public abstract class EfficientDetObjectDetectionModelBase<I> : ObjectDetectionM
 /**
  * Base class for object detection model based on SSD architecture.
  */
-public abstract class SSDObjectDetectionModelBase<I>(protected val metadata: SSDModelMetadata) : ObjectDetectionModelBase<I>() {
+public abstract class SSDLikeModelBase<I>(protected val metadata: SSDLikeModelMetadata) : ObjectDetectionModelBase<I>() {
     override fun convert(output: Map<String, Any>): List<DetectedObject> {
         val boxes = (output[metadata.outputBoxesName] as Array<Array<FloatArray>>)[0]
         val classIndices = (output[metadata.outputClassesName] as Array<FloatArray>)[0]
@@ -92,7 +92,7 @@ public abstract class SSDObjectDetectionModelBase<I>(protected val metadata: SSD
     }
 }
 
-public data class SSDModelMetadata(
+public data class SSDLikeModelMetadata(
     public val outputBoxesName: String,
     public val outputClassesName: String,
     public val outputScoresName: String,
